+++ SmartHomeAPI/ApplicationCore/ApplicationCore/Entities/Temperature.cs
@@ -4,7 +4,7 @@ namespace SmartHomeAPI.ApplicationCore.Entities
 {
     public class Temperature
     {
-        public float Celsius { get; set; }
+        public double Celsius { get; set; }
 
         public DateTime Date { get; set; }
     }
